// src/types/types-1.45.0.ts
export interface VaultExplorerPluginSettings_1_45_0 {
	properties: {
		url: string;
		image: string;
		coverImageFit: string;
		createdDate: string;
		modifiedDate: string;
		custom1: string;
		custom2: string;
		custom3: string;
	};
	filters: {
		search: TSearchFilter;
		sort: TSortFilter;
		custom: TCustomFilter;
	};
	views: {
		dashboard: TDashboardView;
		grid: TGridView;
		list: TListView;
		table: TTableView;
		feed: TFeedView;
		recommended: TRecommendedView;
		related: TRelatedView;
	};
	confirmBeforeDelete: boolean;
	titleWrapping: WordBreak;
	enableClockUpdates: boolean;
	enableFileIcons: boolean;
	loadBodyTags: boolean;
	currentView: TExplorerView | null;
	pageSize: number;
	shouldCollapseFilters: boolean;
	viewOrder: TExplorerView[];
	configDir: string;
	pluginVersion: string | null;
	logLevel: string;
}

interface BaseView {
	isEnabled: boolean;
}

interface TTableView extends BaseView {}

interface TListView extends BaseView {
	showTags: boolean;
}

interface TGridView extends BaseView {
	coverImageSources: CoverImageSource[];
	coverImageFit: CoverImageFit;
	loadSocialMediaImage: boolean;
}

interface CoverImageSource {
	type: CoverImageSourceType;
	isEnabled: boolean;
}

type CoverImageFit = "cover" | "contain";

type CoverImageSourceType =
	| "image-property"
	| "url-property"
	| "frontmatter"
	| "body";

interface TDashboardView extends BaseView {}

type CollapseStyle = "no-new-lines" | "no-extra-new-lines";

interface TFeedView extends BaseView {
	collapseStyle: CollapseStyle;
	removeH1: boolean;
	lineClampSmall: number;
	lineClampMedium: number;
	lineClampLarge: number;
}

interface TRecommendedView extends BaseView {}

interface TRelatedView extends BaseView {}

interface BaseFilter {
	isEnabled: boolean;
}

interface TSearchFilter extends BaseFilter {
	value: string;
}

interface TSortFilter extends BaseFilter {
	value: SortFilterOption;
}

interface TCustomFilter extends BaseFilter {
	selectedGroupId: string;
	groups: TFilterGroup[];
}

type SortFilterOption =
	| "file-name-asc"
	| "file-name-desc"
	| "modified-asc"
	| "modified-desc"
	| "created-asc"
	| "created-desc"
	| "random";

type WordBreak = "normal" | "break-word";

enum TExplorerView {
	DASHBOARD = "dashboard",
	GRID = "grid",
	LIST = "list",
	FEED = "feed",
	TABLE = "table",
	RECOMMENDED = "recommended",
	RELATED = "related",
}
// ...
interface TFilterGroup {
	id: string;
	name: string;
	rules: TFilterRule[];
	isEnabled: boolean;
	isSticky: boolean;
}